Generate the custom linker script from the default.
LD will output the default linker script when invoked with the
'--verbose' option.  This patch causes the build scripts to take that
output and modify it to produce our own custom linker script.

This should fix the issue that people have been having with building
the firmware with older versions of binutils.