Reduce the watchdog timeout, and increase when we have food.
Robots were spending a quite noticeable amount of time driving into
the walls.  This patch reduces that time.

A side effect of reducing this time was that the watchdog would often
be hit when pushing food towards the light.  This is because the
braitenberg code is working to make sure that the bearing is constant,
and the watchdog is looking for a constant bearing...  So, this patch
uses a watchdog timeout 4 times greater when we have food and are
pointing at the light!